requirements.txt
beautifulsoup4
lxml
readability-lxml
tornado
